+++ resources/views/admin/partials/volunteersControl.blade.php
@@ -2,7 +2,7 @@
     <div id="newVolunteerDialog" class="col-sm" style="display:none;">
         <div class="form-group">
             <button id="add-volunteer" class="btn btn-sm btn-primary volunteer-manage-buttons" type="button" onclick="includeVolunteers();">{{ $settings->word('add_volunteer') }}</button>
-            <button id="save-volunteers" class="btn btn-sm btn-success volunteer-manage-buttons" type="button" onclick="saveVolunteers('<?php echo $dataType?>');">{{ $settings->word('save_volunteers') }}</button>
+            <button id="save-volunteers" class="btn btn-sm btn-success volunteer-manage-buttons" type="button" onclick="saveVolunteers('{{ $dataType }}', '{{ $settings->get("phone_number_validation") }}')">{{ $settings->word('save_volunteers') }}</button>
             <button id="include-group" class="btn btn-sm btn-warning volunteer-manage-buttons" type="button" onclick="showGroupsModal();" style="display: none;">{{ $settings->word('include_group') }}</button>
             <button id="volunteers-download-list-csv" class="btn btn-sm btn-secondary volunteer-manage-buttons" type="button" style="display: none;">Volunteer List (CSV)</button>
             <button id="volunteers-download-list-json" class="btn btn-sm btn-secondary volunteer-manage-buttons" type="button" style="display: none;">Volunteer List (JSON)</button>
@@ -170,7 +170,7 @@
                     Volunteer Name:
                 </div>
                 <div class="col-xs-4 volunteer-name-text" style="float: left;">
-                    <input type="text" class="form-control form-control-sm" id="volunteer_name" name="volunteer_name">
+                    <input type="text" class="form-control form-control-sm volunteerName" id="volunteer_name" name="volunteer_name">
                 </div>
                 <div class="col-xs-4 expand-button">
                     <button class="btn btn-sm btn-outline-info volunteerCardBodyToggleButton" type="button" onclick="toggleCardDetails(this);return false;">+</button>
@@ -182,7 +182,7 @@
         </div>
         <div class="card-body volunteerCardBody collapse">
             <div class="form-group form-row form-inline">
-                Phone Number:  <input type="text" class="form-control form-control-sm" id="volunteer_phone_number" name="volunteer_phone_number">
+                Phone Number:  <input type="text" class="form-control form-control-sm volunteerPhoneNumber" id="volunteer_phone_number" name="volunteer_phone_number">
             </div>
             <div class="form-group form-row form-inline">
                 Gender: <select class="form-control form-control-sm" name="volunteer_gender" id="volunteer_gender">