+++ RELEASENOTES.md
@@ -4,6 +4,7 @@
 * Added feature for SPAD playback `spad_option`. 
 * Fix for incorrect calculation of total and missed calls metrics on reports page.
 * Fix for scenario when `sms_combine` is enabled and `sms_ask` is disabled.  [#706]
+* All endpoints can now receive a `callback` querystring parameter that returns JSONP if it's a JSON response endpoint.
 * Yap now requires PHP version 8.1 and higher.
 
 ### 4.2.11 (September 5, 2023)