+++ RELEASENOTES.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-### 4.3.0 (UNRELEASED)
+### 4.3.0 (March 11, 2024)
 * Added feature for SPAD playback `spad_option`.
 * Added the ability to refine the spoken pronunciation of locations. [#917]
 * Added the ability to delete groups. [#680]